#34 - libcore: integrate features
