+++ metrics/regression/current.summary
@@ -1,40 +1,40 @@
 Perf Metrics Summary: core
 --------------------
-core/compile       bytes:     1520     usecs:      40.28
-core/core          bytes:     5808     usecs:     124.80
-core/gcd           bytes:      624     usecs:     119.70
-core/list          bytes:     5296     usecs:     102.86
-core/namespace     bytes:      240     usecs:       3.44
-core/number        bytes:     3600     usecs:      70.20
-core/reader        bytes:     5280     usecs:      89.22
-core/special-form  bytes:     2400     usecs:      56.08
-core/stream        bytes:      480     usecs:      11.28
-core/struct        bytes:      576     usecs:      12.14
-core/symbol        bytes:     1200     usecs:      24.58
-core/vector        bytes:     4456     usecs:      93.04
+core/compile       bytes:     1520     usecs:      32.14
+core/core          bytes:     5808     usecs:     110.04
+core/gcd           bytes:      624     usecs:      89.92
+core/list          bytes:     5296     usecs:      89.24
+core/namespace     bytes:      240     usecs:       2.66
+core/number        bytes:     3600     usecs:      57.68
+core/reader        bytes:     5280     usecs:      74.08
+core/special-form  bytes:     2400     usecs:      43.34
+core/stream        bytes:      480     usecs:       9.02
+core/struct        bytes:      576     usecs:       9.84
+core/symbol        bytes:     1200     usecs:      20.50
+core/vector        bytes:     4456     usecs:      73.66
 
 Perf Metrics Summary: frequent
 --------------------
-frequent/core      bytes:     9888     usecs:     566.36
-frequent/fixnum    bytes:     1680     usecs:      30.54
-frequent/float     bytes:     1200     usecs:      24.86
-frequent/list      bytes:     2160     usecs:      41.80
-frequent/vector    bytes:      240     usecs:       5.00
+frequent/core      bytes:     9888     usecs:     424.04
+frequent/fixnum    bytes:     1680     usecs:      27.54
+frequent/float     bytes:     1200     usecs:      20.22
+frequent/list      bytes:     2160     usecs:      35.52
+frequent/vector    bytes:      240     usecs:       4.14
 
 Perf Metrics Summary: prelude
 --------------------
-prelude/closure    bytes:    20904     usecs:   12928.16
-prelude/compile    bytes:     5512     usecs:    1418.62
-prelude/prelude    bytes:     4592     usecs:     282.58
-prelude/exception  bytes:     7320     usecs:    4779.02
-prelude/fixnum     bytes:     2000     usecs:     187.18
-prelude/format     bytes:     6184     usecs:    6918.46
-prelude/lambda     bytes:    97784     usecs:   64282.74
-prelude/list       bytes:    20864     usecs:    8617.88
-prelude/macro      bytes:    58640     usecs:   43136.42
-prelude/reader     bytes:    15216     usecs:  108960.34
-prelude/stream     bytes:      960     usecs:      70.30
-prelude/string     bytes:     2160     usecs:     553.36
-prelude/type       bytes:     8768     usecs:    5720.22
-prelude/vector     bytes:     9472     usecs:    6063.80
+prelude/closure    bytes:    20904     usecs:    9001.66
+prelude/compile    bytes:     5512     usecs:    1000.76
+prelude/prelude    bytes:     4592     usecs:     250.40
+prelude/exception  bytes:     7320     usecs:    3249.66
+prelude/fixnum     bytes:     2000     usecs:     154.72
+prelude/format     bytes:     6184     usecs:    4507.56
+prelude/lambda     bytes:    97784     usecs:   42548.76
+prelude/list       bytes:    20864     usecs:    5886.30
+prelude/macro      bytes:    58640     usecs:   28304.88
+prelude/reader     bytes:    15216     usecs:   74669.88
+prelude/stream     bytes:      960     usecs:      59.38
+prelude/string     bytes:     2160     usecs:     407.88
+prelude/type       bytes:     8768     usecs:    4044.60
+prelude/vector     bytes:     9472     usecs:    4271.90
