+++ src/libcore/features/std/mod.rs
@@ -5,7 +5,4 @@
 //! system module
 pub struct System {}
 
-pub mod image;
-pub mod process;
-pub mod time;
-pub mod uname;
+pub mod std_;
