+++ src/libcore/system/process.rs
@@ -5,9 +5,9 @@
 use {
     crate::{
         core::{
+            apply::Core as _,
             exception::{self, Condition, Exception},
             frame::Frame,
-            funcall::Core as _,
             mu::Mu,
             types::{Tag, Type},
         },
