+++ src/libcore/types/fixnum.rs
@@ -4,10 +4,10 @@
 //! mu fixnum type
 use crate::{
     core::{
+        apply::Core as _,
         direct::{DirectInfo, DirectTag, DirectType, ExtType},
         exception::{self, Condition, Exception, Result},
         frame::Frame,
-        funcall::Core as _,
         mu::Mu,
         system::Core as _,
         types::{Tag, Type},
